+++ src/main/scala/xitrum/view/Js.scala
@@ -65,11 +65,11 @@ trait Js {
   //----------------------------------------------------------------------------
 
   lazy val jsDefaults = {
-    val validatei18n = if (language == "en") "" else (<script type="text/javascript" src={webJarsUrl(s"jquery-validation/1.13.0/localization/messages_$language.js")}></script>)
+    val validatei18n = if (language == "en") "" else (<script type="text/javascript" src={webJarsUrl(s"jquery-validation/1.13.1/localization/messages_$language.js")}></script>)
     <xml:group>
       <script type="text/javascript" src={webJarsUrl("jquery/2.1.1",              "jquery.js",             "jquery.min.js")}></script>
-      <script type="text/javascript" src={webJarsUrl("jquery-validation/1.13.0",  "jquery.validate.js",    "jquery.validate.min.js")}></script>
-      <script type="text/javascript" src={webJarsUrl("jquery-validation/1.13.0",  "additional-methods.js", "additional-methods.min.js")}></script>
+      <script type="text/javascript" src={webJarsUrl("jquery-validation/1.13.1",  "jquery.validate.js",    "jquery.validate.min.js")}></script>
+      <script type="text/javascript" src={webJarsUrl("jquery-validation/1.13.1",  "additional-methods.js", "additional-methods.min.js")}></script>
       {validatei18n}
       <script type="text/javascript" src={webJarsUrl("sockjs-client/0.3.4",       "sockjs.js",             "sockjs.min.js")}></script>
       <script type="text/javascript" src={url[xitrum.js]}></script>